Move timezone_getter to conftest
