+++ tests/test_dates.py
@@ -16,39 +16,9 @@ from datetime import date, datetime, time, timedelta
 import freezegun
 import pytest
 
-# for tests it can be useful to have both zoneinfo and pytz available
-try:
-    import zoneinfo
-except ModuleNotFoundError:
-    try:
-        from backports import zoneinfo
-    except ImportError:
-        zoneinfo = None
-
-try:
-    import pytz
-except ModuleNotFoundError:
-    pytz = None
-
 from babel import dates, Locale
-from babel.dates import NO_INHERITANCE_MARKER, _localize, _get_tz_name, LOCALTZ
+from babel.dates import NO_INHERITANCE_MARKER, _localize
 from babel.util import FixedOffsetTimezone
-
-
-@pytest.fixture(params=["pytz.timezone", "zoneinfo.ZoneInfo"], scope="package")
-def timezone_getter(request):
-    if request.param == "pytz.timezone":
-        if pytz:
-            return pytz.timezone
-        else:
-            pytest.skip("pytz not available")
-    elif request.param == "zoneinfo.ZoneInfo":
-        if zoneinfo:
-            return zoneinfo.ZoneInfo
-        else:
-            pytest.skip("zoneinfo not available")
-    else:
-        raise NotImplementedError
 
 
 class DateTimeFormatTestCase:
